adding settings and vision api

+++ app/routes/chat_route.py
@@ -1,7 +1,12 @@
 from flask import Blueprint, request, Response
 from flask_socketio import join_room
+from pymongo import MongoClient
+import certifi
+import os
+import json
 from app import socketio
 from app.services.ChatService import ChatService
+from app.services.UserService import UserService
 from app.agents.BossAgent import BossAgent
 from dotenv import load_dotenv
 
@@ -9,9 +14,13 @@ load_dotenv()
 
 chat_bp = Blueprint('chat', __name__)
 
-headers = {"Access-Control-Allow-Origin": "*"}
+mongo_uri = os.getenv('MONGO_URI')
+client = MongoClient(mongo_uri, tlsCAFile=certifi.where())
+
+db = client['paxxium']
 
 chat_service = ChatService('paxxium')
+user_service = UserService(db)
 
 @socketio.on('connect')
 def handle_connect():
@@ -37,7 +46,7 @@ def chat(subpath):
 
     if request.method == 'GET' and subpath == '':
         user_id = request.headers.get('userId')
-        return chat_service.get_all_chats(user_id), 200, headers
+        return chat_service.get_all_chats(user_id), 200
     
     if request.method == 'POST' and subpath == '':
         data = request.get_json()
@@ -59,19 +68,19 @@ def chat(subpath):
             'chatConstants': data.get('chatConstants'),
             'useProfileData': data.get('useProfileData'),
             'is_open': True
-        }, 200, headers
+        }, 200
 
     if request.method == 'DELETE' and subpath == '':
         chat_id = request.get_json()['chatId']
         chat_service.delete_chat(chat_id)
-        return 'Conversation deleted', 200, headers
+        return 'Conversation deleted', 200
     
     if subpath == 'update_visibility':
         data = request.get_json()
         chat_id = data['chatId']
         is_open = data['is_open']
         chat_service.update_visibility(chat_id, is_open)
-        return ('Chat visibility updated', 200, headers)
+        return ('Chat visibility updated', 200)
 
     if subpath == 'update_settings':
         data = request.get_json()
@@ -83,7 +92,7 @@ def chat(subpath):
         chat_constants = data.get('chat_constants')
         chat_service.update_settings(chat_id, chat_name, agent_model, system_prompt, chat_constants, use_profile_data)
 
-        return ('Chat settings updated', 200, headers)
+        return ('Chat settings updated', 200)
 
 @socketio.on('join_room')
 def handle_join_room(data):
@@ -93,27 +102,51 @@ def handle_join_room(data):
 
 @socketio.on('chat_request')
 def handle_chat_message(data):
-    save_to_db = data.get('saveToDb', True)
-    create_vector_pipeline = data.get('createVectorPipeline', False)
-    boss_agent = BossAgent()
+    uid = data['userId']
+    
+    
+    chat_settings = data.get('chatSettings', None)
+    if chat_settings:
+        chat_constants = chat_settings.get('chat_constants')
+        use_profile_data = chat_settings.get('use_profile_data', None)
+        system_prompt = chat_settings.get('system_prompt')
+        if use_profile_data:
+            user_analysis = user_service.get_user_analysis(uid)
+        boss_agent = BossAgent(chat_constants=chat_constants, system_prompt=system_prompt, user_analysis=user_analysis)
+    else:
+        boss_agent = BossAgent()
+    
     chat_service_with_db = ChatService(db_name=data['dbName'])  
     user_message = data['userMessage']['content']
     chat_id = data['chatId']
 
+    create_vector_pipeline = data.get('createVectorPipeline', False)
     if create_vector_pipeline:
         query_pipeline = boss_agent.create_vector_pipeline(user_message, data['projectId'])
         results = chat_service_with_db.query_snapshots(query_pipeline)
-        system_message = boss_agent.prepare_vector_response(results)
+        system_message = boss_agent.prepare_vector_response(results, system_prompt)
     else:
         system_message = None
 
+    save_to_db = data.get('saveToDb', True)
     if save_to_db:
         chat_service.create_message(chat_id, 'user', user_message)
 
-    boss_agent.process_message(data['chatHistory'], chat_id, user_message, system_message)
+    def save_agent_message(chat_id, message):
+        chat_service.create_message(chat_id, 'agent', message)
 
-@chat_bp.route('/messages', methods=['DELETE'])
-def handle_delete_all_messages():
-    chat_id = request.json.get('chatId')
-    chat_service.delete_all_messages(chat_id)
-    return 'Memory Cleared', 200, headers
+    boss_agent.process_message(data['chatHistory'], chat_id, user_message, system_message, save_agent_message if save_to_db else None)
+
+@chat_bp.route('/messages', defaults={'subpath': ''}, methods=['DELETE'])
+@chat_bp.route('/messages/<path:subpath>', methods=['DELETE'])
+def handle_delete_all_messages(subpath):
+    if request.method == 'DELETE' and subpath == '':
+        chat_id = request.json.get('chatId')
+        chat_service.delete_all_messages(chat_id)
+        return 'Memory Cleared', 200
+    
+    if subpath == 'utils':
+        uid = request.headers.get('userId')
+        file = request.files['image']
+        file_url = user_service.upload_file_to_firebase_storage(file, uid)
+        return (json.dumps({'fileUrl': file_url}), 200)
